Consolidated explosives permit interfaces
