+++ services/common/src/interfaces/permits/index.ts
@@ -3,7 +3,6 @@ export * from "./createPermitConditionPayload.interface";
 export * from "./createPermitPayload.interface";
 export * from "./createPermitSiteProperties.interface";
 export * from "./explosivesPermit.interface";
-export * from "./magazine.interface";
 export * from "./permit.interface";
 export * from "./permitAmendment.interface";
 export * from "./permitAmendmentDocument.interface";
@@ -14,7 +13,5 @@ export * from "./standardPermitCondition.interface";
 export * from "./permitParty.interface";
 export * from "./permitPartyRelationship.interface";
 export * from "./updatePermitAmendmentPayload.interface";
-export * from "./PermitAmendmentMineDocument.interface";
 export * from "./patchPermitNumber.interface";
 export * from "./draftPermitAmendment.interface";
-export * from "./explosivesPermitAmendment.interface";
