--- services/common/src/interfaces/permits/magazine.interface.ts
@@ -1,17 +1,0 @@
-export interface IMagazine {
-  explosives_permit_magazine_id: number;
-  explosives_permit_id: number;
-  explosives_permit_magazine_type_code: string;
-  type_no: string;
-  tag_no: string;
-  construction: string;
-  latitude: number;
-  longitude: number;
-  length: number;
-  width: number;
-  height: number;
-  quantity: number;
-  distance_road: number;
-  distance_dwelling: number;
-  detonator_type: string;
-}
