+++ services/core-web/common/actionCreators/explosivesPermitActionCreator.ts
@@ -1,6 +1,6 @@
 import { notification } from "antd";
 import { showLoading, hideLoading } from "react-redux-loading-bar";
-import { ENVIRONMENT } from "@mds/common";
+import { ENVIRONMENT, IExplosivesPermit } from "@mds/common";
 import { request, success, error } from "../actions/genericActions";
 import * as reducerTypes from "../constants/reducerTypes";
 import * as explosivesPermitActions from "../actions/explosivesPermitActions";
@@ -8,7 +8,6 @@ import * as String from "../constants/strings";
 import * as API from "../constants/API";
 import { createRequestHeader } from "../utils/RequestHeaders";
 import CustomAxios from "../customAxios";
-import { IExplosivesPermit } from "@mds/common";
 import { AppThunk } from "@/store/appThunk.type";
 import { AxiosResponse } from "axios";
